LayoutTests/imported/w3c/web-platform-tests/html/rendering/widgets/field-sizing-input-number-expected.txt



FAIL number: Empty value assert_equals: expected 19 but got 23
PASS number: Auto width and auto height with a long text
PASS number: Explicit width and heigth
FAIL number: Explicit width and auto height assert_equals: expected 19 but got 23
PASS number: Explicit height and auto width
PASS number: Text caret is taller than the placeholder
FAIL number: Text caret is shorter than the placeholder assert_less_than: expected a number less than 36 but got 36
FAIL number: Update field-sizing property dynamically assert_equals: expected 19 but got 23
